edit some modules
