--- js/comment-length.js
@@ -1,4 +1,0 @@
-const calculateLengthComment = (commentText, maxLengthComment) =>
-  commentText.length <= maxLengthComment;
-
-export {calculateLengthComment};
