+++ js/utils.js
@@ -1,3 +1,5 @@
+const MAX_QUANTITY_SYMBOL = 140;
+
 // Функция для получения случайного числа из указанного интервала.
 // Решила использовать функцию Кекса :) Она более краткая и понятная.
 const getRandomPositiveInteger = (a, b) => {
@@ -12,3 +14,11 @@ const getRandomArrayElement = (elements) =>
   elements[getRandomPositiveInteger(0, elements.length - 1)];
 
 export {getRandomPositiveInteger, getRandomArrayElement};
+
+// Вычисляем длину строки
+const calculateLengthComment = (commentText, maxLengthComment) =>
+  commentText.length <= maxLengthComment;
+
+export {calculateLengthComment};
+
+calculateLengthComment('Какой-то комментарий пользователя' , MAX_QUANTITY_SYMBOL);
